GIT_SILENT Fix build for msvs 2015 (float/double types)

+++ examples/KReportExampleDataSource.h
@@ -62,8 +62,8 @@ private:
         QString project;
         QString country;
         QString mobile;
-        float lat;
-        float lon;
+        qreal lat;
+        qreal lon;
         QString code;
         bool projectLead;
     };
